+++ test-parts/part-g-controller_files/controller-main.js
@@ -20,9 +20,9 @@ function coordinateController()
 		cRioFactoriesFile.callTestRemoteIoFactories();
 		cRioIndexMainFile.callTestRemoteIoIndexMain();
 		cRioIndexNodeListFile.callTestRemoteIoIndexNodeList();
+		cRioIndexNodeRegisterFile.callTestRemoteIoIndexNodeReg();
 		*/
 		
-		cRioIndexNodeRegisterFile.callTestRemoteIoIndexNodeReg();
 		cRioIndexNodeRegisterInvalid.callTestRemoteIoIndexRegisterInvalid();
 		
 		/*